Moved optimized_driver to lcd_driver folder and rearranged indentation in optimized_driver
