+++ Lab2/src/hi/tasks/lcd_draw_task.cpp
@@ -18,7 +18,6 @@
  **/
 
 #include "hi/tasks/lcd_draw_task.hh"
-#include "hd/optimized_driver.hh"
 
 LcdDrawTask::LcdDrawTask(void)
 {
